Rename index.js files placed in various subfolders to be prefixed with name of enclosing subdirectory; Exclude test/ and __mocks__/ folders from ESLint config; Exclude src/frontend/assets folder from being passed to Prettier;
