+++ package.json
@@ -21,11 +21,11 @@
   "scripts": {
     "test:backend": "jest --watch --env=node",
     "test:debug": "node --inspect node_modules/jest/bin/jest.js --runInBand",
-    "prettier": "prettier --write src/ !src/database/data",
+    "prettier": "prettier --write src/ !src/frontend/assets !src/database/data",
     "prettier:backend": "prettier --write src/middleware/ src/database/ !src/database/data",
     "eslint:backend": "eslint src/middleware/** src/database/**",
     "start": "set \"NODE_ONLY=false\" && webpack-dev-server",
-    "serve": "set \"NODE_ONLY=true\" && nodemon --exec \"npm run eslint:backend && npm run prettier:backend && node src/middleware/index.js\"",
+    "serve": "set \"NODE_ONLY=true\" && nodemon --exec \"npm run eslint:backend && npm run prettier:backend && node src/middleware/middleware-index.js\"",
     "prebuild": "npm run prettier",
     "build": "webpack"
   },
