+++ test/database/utils/paginateItemsFromDB.spec.js
@@ -1,5 +1,5 @@
 // TODO: create kind of symlinks to test/ folder to avoid using relative paths
-const { findAssociatedSrcModulePath } = require('../../index');
+const { findAssociatedSrcModulePath } = require('../../test-index');
 const getPaginatedItems = require(findAssociatedSrcModulePath());
 
 describe('#paginateItemsFromDB', () => {
